refactoring tables
